Indexing, Callbacks from tiles
All test movement and callbacks tested and functioning as intended. Fixed bug to change the target of the callback on iteration.

Added const defs file to include for no magic numbers, changes occur in one file. Since size of grid is known for development and isn't something we need to set dynamically.

BUG FIXES: Fixed memory leak in grid delete that while the tiles were freed, the array allocated for tiles itself was not.